Address Peiyu's feedback.
